+++ CHANGELOG.md
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Edit on Github` button now hidden for plugins.
 - Doctests now run via unittests
 - Every class and function can now be imported from the top level module
+- `op` attempts to create `Definition`s for each argument if an `inputs` are not
+  given.
 ### Fixed
 - Race condition in `MemoryRedundancyChecker` when more than 4 possible
   parameter sets for an operation.